sam-max-cases/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-redis==5.4.0
celery==5.4.0
redis==5.1.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyyaml==6.0.2
requests==2.32.3
Pillow==10.4.0
jinja2==3.1.4
python-magic==0.4.27
cryptography==43.0.1
pickle5==0.0.12
reportlab==4.2.5
